// SKYNET server/GC/730/generated/cs2.ts
export interface ProtoDescriptor<TMessage> {
    readonly name: string;
}

export interface GcRoute<TRequest, TResponse> {
    readonly requestId: number;
    readonly request: ProtoDescriptor<TRequest>;
    readonly responseId: number;
    readonly response: ProtoDescriptor<TResponse>;
}

export interface EmptyRequest {
}

export type MessageId = number;
// ...
export interface CMsgGCCStrike15v2ClientGCRankUpdate {
    readonly rankings?: PlayerRankingInfo[];
}
// ...
export interface CMsgGCCStrike15v2MatchEndRunRewardDrops {
    readonly serverinfo?: CMsgGCCStrike15v2MatchmakingServerReservationResponse;
    readonly matchEndQuestData?: CMsgGCServerQuestUpdateData;
}
// ...
export interface CMsgGCCStrike15v2MatchmakingGC2ClientReserve {
    readonly serverid?: bigint;
    readonly directUdpIp?: number;
    readonly directUdpPort?: number;
    readonly reservationid?: bigint;
    readonly reservation?: CMsgGCCStrike15v2MatchmakingGC2ServerReserve;
    readonly map?: string;
    readonly serverAddress?: string;
    readonly gsPing?: DataCenterPing;
    readonly gsLocationId?: number;
}

export interface CMsgGCCStrike15v2MatchmakingGC2ClientUpdate {
    readonly matchmaking?: number;
    readonly waitingAccountIdSessions?: number[];
    readonly error?: string;
    readonly ongoingmatchAccountIdSessions?: number[];
    readonly globalStats?: GlobalStatistics;
    readonly failpingAccountIdSessions?: number[];
    readonly penaltyAccountIdSessions?: number[];
    readonly failreadyAccountIdSessions?: number[];
    readonly vacbannedAccountIdSessions?: number[];
    readonly serverIpaddressMask?: IpAddressMask;
    readonly notes?: CMsgGCCStrike15v2MatchmakingGC2ClientUpdateNote[];
    readonly penaltyAccountIdSessionsGreen?: number[];
    readonly insufficientlevelSessions?: number[];
    readonly vsncheckAccountIdSessions?: number[];
    readonly launcherMismatchSessions?: number[];
    readonly insecureAccountIdSessions?: number[];
}

export interface CMsgGCCStrike15v2MatchmakingGC2ServerReserve {
    readonly accountIds?: number[];
    readonly gameType?: number;
    readonly matchId?: bigint;
    readonly serverVersion?: number;
    readonly rankings?: PlayerRankingInfo[];
    readonly encryptionKey?: bigint;
    readonly encryptionKeyPub?: bigint;
    readonly partyIds?: number[];
    readonly whitelist?: IpAddressMask[];
    readonly tvMasterSteamid?: bigint;
    readonly tournamentEvent?: TournamentEvent;
    readonly tournamentTeams?: TournamentTeam[];
    readonly tournamentCastersAccountIds?: number[];
    readonly tvRelaySteamid?: bigint;
    readonly preMatchData?: CPreMatchInfoData;
    readonly tvControl?: number;
    readonly flags?: number;
    readonly opVarValues?: OperationalVarValue[];
    readonly socacheControl?: number;
    readonly teammateColors?: number[];
    readonly matchIdAdditional?: number;
}

export interface CMsgGCCStrike15v2MatchmakingServerReservationResponse {
    readonly reservationid?: bigint;
    readonly reservation?: CMsgGCCStrike15v2MatchmakingGC2ServerReserve;
    readonly map?: string;
    readonly gcReservationSent?: bigint;
    readonly serverVersion?: number;
    readonly tvInfo?: ServerHltvInfo;
    readonly rewardPlayerAccounts?: number[];
    readonly idlePlayerAccounts?: number[];
    readonly rewardItemAttrDefIdx?: number;
    readonly rewardItemAttrValue?: number;
    readonly rewardItemAttrRewardIdx?: number;
    readonly rewardDropList?: number;
    readonly tournamentTag?: string;
    readonly legacySteamdatagramPort?: number;
    readonly testToken?: number;
    readonly flags?: number;
    readonly steamdatagramRouting?: number;
    readonly systemLoad?: number;
    readonly cpusOnline?: number;
}

export interface CMsgGCCStrike15v2MatchmakingStart {
    readonly accountIds?: number[];
    readonly gameType?: number;
    readonly ticketData?: string;
    readonly clientVersion?: number;
    readonly tournamentMatch?: TournamentMatchSetup;
    readonly primeOnly?: boolean;
    readonly tvControl?: number;
    readonly lobbyId?: bigint;
}

export interface CMsgGCCStrike15v2MatchmakingStop {
    readonly abandon?: number;
}
// ...
export interface PlayerRankingInfo {
    readonly accountId?: number;
    readonly rankId?: number;
    readonly wins?: number;
    readonly rankChange?: number;
    readonly rankTypeId?: number;
    readonly tvControl?: number;
    readonly rankWindowStats?: bigint;
    readonly leaderboardName?: string;
    readonly rankIfWin?: number;
    readonly rankIfLose?: number;
    readonly rankIfTie?: number;
    readonly perMapRank?: PlayerRankingInfo_PerMapRank[];
    readonly leaderboardNameStatus?: number;
    readonly highestRank?: number;
    readonly rankExpiry?: number;
}

export const Msg: {
    SOCacheSubscribed: MessageId;
    SOMultipleObjects: MessageId;
    SOCacheSubscriptionRefresh: MessageId;
    GCSetItemPosition: MessageId;
    GCVerifyCacheSubscription: MessageId;
    GCSetItemPositions: MessageId;
    GCApplySticker: MessageId;
    GCItemAcknowledged: MessageId;
    GCItemCustomizationNotification: MessageId;
    GCStoreGetUserData: MessageId;
    GCStoreGetUserDataResponse: MessageId;
    GCAdjustEquipSlotsManual: MessageId;
    GCAdjustEquipSlotsShuffle: MessageId;
    GCOpenCrate: MessageId;
    GCClientWelcome: MessageId;
    GCGameServerWelcome: MessageId;
    GCClientHello: MessageId;
    GCGameServerHello: MessageId;
    GCClientConnectionStatus: MessageId;
    MatchmakingStart: MessageId;
    MatchmakingStop: MessageId;
    MatchmakingClient2ServerPing: MessageId;
    MatchmakingGC2ClientUpdate: MessageId;
    MatchmakingGC2ServerReserve: MessageId;
    MatchmakingServerReservationResponse: MessageId;
    MatchmakingGC2ClientReserve: MessageId;
    MatchmakingClient2GCHello: MessageId;
    MatchmakingGC2ClientHello: MessageId;
    ClientRequestPlayersProfile: MessageId;
    PlayersProfile: MessageId;
    MatchEndRunRewardDrops: MessageId;
    MatchList: MessageId;
    MatchListRequestRecentUserGames: MessageId;
    MatchListRequestTournamentGames: MessageId;
    Server2GCClientValidate: MessageId;
    MatchListRequestTournamentPredictions: MessageId;
    AcknowledgePenalty: MessageId;
    ClientPlayerDecalSign: MessageId;
    PartySearch: MessageId;
    AccountRequestCoPlays: MessageId;
    ClientGCRankUpdate: MessageId;
    GetEventFavoritesRequest: MessageId;
    GetEventFavoritesResponse: MessageId;
    PremierSeasonSummary: MessageId;
    RequestRecurringMissionSchedule: MessageId;
    RecurringMissionSchema: MessageId;
    VolatileShopSubscribe: MessageId;
} = {
    SOCacheSubscribed: 24,
    SOMultipleObjects: 26,
    SOCacheSubscriptionRefresh: 28,
    GCSetItemPosition: 1001,
    GCVerifyCacheSubscription: 1005,
    GCSetItemPositions: 1077,
    GCApplySticker: 1086,
    GCItemAcknowledged: 1087,
    GCItemCustomizationNotification: 1090,
    GCStoreGetUserData: 2500,
    GCStoreGetUserDataResponse: 2501,
    GCAdjustEquipSlotsManual: 2531,
    GCAdjustEquipSlotsShuffle: 2532,
    GCOpenCrate: 2534,
    GCClientWelcome: 4004,
    GCGameServerWelcome: 4005,
    GCClientHello: 4006,
    GCGameServerHello: 4007,
    GCClientConnectionStatus: 4009,
    MatchmakingStart: 9101,
    MatchmakingStop: 9102,
    MatchmakingClient2ServerPing: 9103,
    MatchmakingGC2ClientUpdate: 9104,
    MatchmakingGC2ServerReserve: 9105,
    MatchmakingServerReservationResponse: 9106,
    MatchmakingGC2ClientReserve: 9107,
    MatchmakingClient2GCHello: 9109,
    MatchmakingGC2ClientHello: 9110,
    ClientRequestPlayersProfile: 9127,
    PlayersProfile: 9128,
    MatchEndRunRewardDrops: 9136,
    MatchList: 9139,
    MatchListRequestRecentUserGames: 9141,
    MatchListRequestTournamentGames: 9146,
    Server2GCClientValidate: 9153,
    MatchListRequestTournamentPredictions: 9160,
    AcknowledgePenalty: 9171,
    ClientPlayerDecalSign: 9185,
    PartySearch: 9191,
    AccountRequestCoPlays: 9193,
    ClientGCRankUpdate: 9194,
    GetEventFavoritesRequest: 9201,
    GetEventFavoritesResponse: 9203,
    PremierSeasonSummary: 9224,
    RequestRecurringMissionSchedule: 9225,
    RecurringMissionSchema: 9226,
    VolatileShopSubscribe: 9228,
};

export const Proto = {
// ...
    CMsgGCCStrike15v2ClientGCRankUpdate: { name: "Cs2Proto.CMsgGCCStrike15v2ClientGCRankUpdate" } as ProtoDescriptor<CMsgGCCStrike15v2ClientGCRankUpdate>,
// ...
    CMsgGCCStrike15v2MatchEndRunRewardDrops: { name: "Cs2Proto.CMsgGCCStrike15v2MatchEndRunRewardDrops" } as ProtoDescriptor<CMsgGCCStrike15v2MatchEndRunRewardDrops>,
// ...
    CMsgGCCStrike15v2MatchmakingGC2ClientReserve: { name: "Cs2Proto.CMsgGCCStrike15v2MatchmakingGC2ClientReserve" } as ProtoDescriptor<CMsgGCCStrike15v2MatchmakingGC2ClientReserve>,
    CMsgGCCStrike15v2MatchmakingGC2ClientUpdate: { name: "Cs2Proto.CMsgGCCStrike15v2MatchmakingGC2ClientUpdate" } as ProtoDescriptor<CMsgGCCStrike15v2MatchmakingGC2ClientUpdate>,
    CMsgGCCStrike15v2MatchmakingGC2ServerReserve: { name: "Cs2Proto.CMsgGCCStrike15v2MatchmakingGC2ServerReserve" } as ProtoDescriptor<CMsgGCCStrike15v2MatchmakingGC2ServerReserve>,
    CMsgGCCStrike15v2MatchmakingServerReservationResponse: { name: "Cs2Proto.CMsgGCCStrike15v2MatchmakingServerReservationResponse" } as ProtoDescriptor<CMsgGCCStrike15v2MatchmakingServerReservationResponse>,
    CMsgGCCStrike15v2MatchmakingStart: { name: "Cs2Proto.CMsgGCCStrike15v2MatchmakingStart" } as ProtoDescriptor<CMsgGCCStrike15v2MatchmakingStart>,
    CMsgGCCStrike15v2MatchmakingStop: { name: "Cs2Proto.CMsgGCCStrike15v2MatchmakingStop" } as ProtoDescriptor<CMsgGCCStrike15v2MatchmakingStop>,
// ...
} as const;